Add internal/interface file split.
