+++ lib/include/defines.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -24,3 +26,17 @@
     fprintf(stderr, "ERROR in line %d: %s\n", __LINE__, m); \
     goto bail; \
 }
+
+typedef struct
+{
+	GLint ESShaderProgram;
+	GLbyte* FShaderSource; //the shader programs are specified as strings and are loaded depending on the program (maybe should be binary blobs in the final impl)
+	GLbyte* VShaderSource; //ditto
+	EGLDisplay display;
+	EGLConfig config;
+	EGLContext context;
+    EGLSurface surface;
+    struct gbm_device* gbm;
+    struct gbm_surface* gbm_surface;
+    int32_t gbd_fd;
+} GLHelper;
